- added filter icon and toggle functionality - did some documentation

+++ applications/mne_browse_raw_qt/Windows/scalewindow.h
@@ -137,7 +137,7 @@ signals:
     /**
     * updateDataTableViews is emmited whenever a connected view spin box value changed
     */
-    void scalingViewValueChanged(double);
+    void scalingViewValueChanged(int);
 
 private:
     Ui::ScaleWindow *ui;
